PIM-7132: Fix "url too long" when doing a mass edit
Context:
When doing a mass-edit in 1.6 and 1.7 (mass-edit has been reworked in
2.0 <3 @juliensnz), the ids of the products selected in the
datagrid would be passed on all the mass-edit steps through the
URL parameter "values".

When submitting the form to create the mass-edit operation, the form
would be submitted via a POST http verb, but the selected Ids would still be
passed on to the controller using the URI parameter.

This submit action would sometimes not execute because the URL would be
too long.

Solution:
Instead of passing on the selected ids through an url parameter, we used
a hidden input in the mass-edit form that would "store" the product ids
in the markup.

When submitting the form to run the mass-edit, we just now need to fetch
the product ids in the right form field: this work is mainly done in the
'Pim\Bundle\DataGridBundle\Extension\MassAction\MassActionDispatcher::prepareMassActionParameters`

As this class is also used for the Quick-exports, we need to keep the
old way of getting the value (through a URL parameter) in order to not
break the quick-exports.